pyproject.toml
[project]
name = "mmtqa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "beautifulsoup4>=4.14.2",
    "datasets<3",
    "datasketch>=1.6.5",
    "google-generativeai>=0.8.5",
    "html5lib>=1.1",
    "huggingface-hub>=0.35.3",
    "nltk>=3.9.2",
    "openai>=2.2.0",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.7",
    "pyarrow>=21.0.0",
    "pydantic>=2.11.10",
    "pymupdf>=1.26.4",
    "requests>=2.32.5",
    "termcolor>=3.1.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.0",
]
